requirements.txt
Django==4.2.8
djangorestframework==3.14.0
mysqlclient==2.2.0
requests==2.32.3
python-dotenv==1.0.1
